Feat: 게시글 상세 api, 게시판에서 글 눌러서 연결 (아직 제대로 작동 안함), CommentResDto, BulletinResDto, BulletinDetailScreen stateless
MainNavigation rebase 해결 예정

+++ core/domain/src/main/java/kr/co/domain/entity/BulletinEntity.kt
@@ -11,7 +11,7 @@ data class BulletinEntity(
     val bulletinCategory: BulletinCategory,
     val createdTime: String,
     val comments: List<CommentEntity>,
-    val bookmarkedCount: Long,
+    val bookmarkedCount: Int,
 ) {
     companion object {
         fun dummy(idx: Int = 0) = BulletinEntity(
@@ -25,7 +25,7 @@ data class BulletinEntity(
             bulletinCategory = BulletinCategory.entries[(idx) % BulletinCategory.entries.size],
             createdTime = "createdTime${idx}",
             comments = emptyList(),
-            bookmarkedCount = 220 + (idx).toLong(),
+            bookmarkedCount = 220 + idx,
         )
     }
